requirements.txt
nltk
TextBlob
newspaper3k
lxml_html_clean
